+++ ContactsControl/Repositories/Contact/IContactRepository.cs
@@ -5,9 +5,9 @@ namespace ContactsControl.Repositories
     public interface IContactRepository
     {
         List<ContactModel> GetAll();
-        ContactModel Get(int id);
+        ContactModel Get(long id);
         ContactModel Create(ContactModel contact);
         ContactModel Update(ContactModel contact);
-        bool Delete(int id);
+        bool Delete(long id);
     }
 }
